Make `Address` implement `Sink`
This allows us to delete the entire `sink.rs` file.
