+++ examples/interleaved_messages.rs
@@ -22,9 +22,9 @@ impl Handler<Hello> for ActorA {
 
     async fn handle(&mut self, _: Hello, ctx: &mut Context<Self>) {
         println!("ActorA: Hello");
-        ctx.handle_while(self, self.actor_b.send(Hello))
-            .await
-            .unwrap();
+        let future = self.actor_b.send(Hello);
+
+        ctx.handle_while(self, future).await.unwrap();
     }
 }
 
@@ -44,7 +44,9 @@ impl Handler<Initialized> for ActorB {
     async fn handle(&mut self, m: Initialized, ctx: &mut Context<Self>) {
         println!("ActorB: Initialized");
         let actor_a = m.0;
-        ctx.handle_while(self, actor_a.send(Hello)).await.unwrap();
+        let future = actor_a.send(Hello);
+
+        ctx.handle_while(self, future).await.unwrap();
     }
 }
